feat: create abstarct menu service
